add up/down cursor movement, basic backspace + next-line wrapping

+++ structs.h
@@ -1,10 +1,12 @@
 #ifndef STRUCTS_H_
 #define STRUCTS_H_
 
+#include <stdint.h>
+
 // structures for buffer and line editing
 typedef struct {
     char* content;
-    int allocated;
+    uint64_t allocated;
 } Buffer;
 
 struct Line {
